Refactor MCP DTO schema model helpers
